po 数据类型修改

+++ legou-order/legou-order-interface/src/main/java/com/hh/legou/order/po/Order.java
@@ -28,31 +28,31 @@ public class Order extends BaseEntity {
      * 数量合计
      */
     @TableField("total_num_")
-    private Integer totalNum;
+    private Long totalNum;
 
     /**
      * 金额合计
      */
     @TableField("total_money_")
-    private Integer totalMoney;
+    private Long totalMoney;
 
     /**
      * 优惠金额
      */
     @TableField("pre_money_")
-    private Integer preMoney;
+    private Long preMoney;
 
     /**
      * 邮费
      */
     @TableField("post_fee_")
-    private Integer postFee;
+    private Long postFee;
 
     /**
      * 实付金额
      */
     @TableField("pay_money_")
-    private Integer payMoney;
+    private Long payMoney;
 
     /**
      * 支付类型，1、在线支付、0 货到付款
